# Conflicts:
#	src/main/java/com/revature/p1/app/services/BookService.java
#	src/main/java/com/revature/p1/app/web/util/ContextLoaderListener.java
#	src/main/java/com/revature/p1/orm/data/QueryBuilder.java

+++ src/main/java/com/revature/p1/app/services/BookService.java
@@ -2,7 +2,6 @@ package com.revature.p1.app.services;
 
 import com.revature.p1.app.daos.BookDAO;
 import com.revature.p1.app.models.Book;
-// TODO - Replace SqlTable with something more sensical
 
 import java.util.List;
 
@@ -21,18 +20,31 @@ public class BookService {
     }
 
     public List<Book> getBookById(String uuid) {
+        if(uuid.length() != 36) return null;
         return bookDAO.getById(uuid);
     }
 
     public boolean insertBook(Book book) {
+        if(!isBookValid(book)) return false;
         return bookDAO.insert(book);
     }
 
     public boolean updateBook(Book book) {
+        if(!isBookValid(book)) return false;
         return bookDAO.update(book);
     }
 
     public boolean deleteBook(Book book) {
+        if(!isBookValid(book)) return false;
         return bookDAO.delete(book);
+    }
+
+    public boolean isBookValid(Book book) {
+        if (book == null) return false;
+        if (book.getTitle() == null || book.getTitle().trim().equals("")) return false;
+        if (book.getAuthor() == null || book.getAuthor().trim().equals("")) return false;
+        if (book.getPageCount() == null || book.getPageCount() < 0) return false;
+
+        return true;
     }
 }
